Fix CI flakiness and static analysis regressions
